func: wrapper added but not implemented
